requirements.txt
dnspython==2.8.0
Faker==37.12.0
psycopg2-binary==2.9.11
pymongo==4.15.3
tzdata==2025.2
